Update to previous proposed solution correction
In my last PR I suggested a new solution to 3.6.1 question 6 to get the "halo" effect on the points.  In my PR I mentioned that I was unhappy with how "size" was functioning.  I just realized size should be outside of `aes` for the second `geom-point()`. This PR corrects that problem.